fix : move getTrash from folder to file
