+++ src/controllers/fileController.js
@@ -1,4 +1,10 @@
-import { uploadFile2, downloadFile2, findFile2, updateFile2 } from '../services/fileService.js';
+import {
+  uploadFile2,
+  downloadFile2,
+  findFile2,
+  updateFile2,
+  getTrash2,
+} from '../services/fileService.js';
 
 export const uploadFile = async (req, res, next) => {
   try {
@@ -44,3 +50,14 @@ export const updateFile = async (req, res, next) => {
 //     return res.status(500).json(err);
 //   }
 // };
+
+// 김태영 file 삭제, 휴지통 조회, 복원
+
+export const getTrash = async (req, res, next) => {
+  try {
+    let rows = await getTrash2(req);
+    return res.json(rows);
+  } catch (err) {
+    return res.status(500).json(err);
+  }
+};
