+++ src/routes/folderRoute.js
@@ -5,7 +5,6 @@ import {
   updateFolder,
   deleteFolder,
   restoreFolder,
-  getTrash,
 } from '../controllers/folderController.js';
 
 const router = express.Router();
@@ -17,5 +16,4 @@ router.patch('/move', moveFolder); // 폴더 이동
 router.delete('/', deleteFolder);
 router.patch('/trash', restoreFolder);
 
-router.get('/trash', getTrash); // 쓰레기통에 있는 폴더 & 파일 조회
 export default router;
