+++ src/services/folderService.js
@@ -797,75 +797,3 @@ export const restoreFolder2 = async (req) => {
     throw Error(err);
   }
 };
-
-export const getTrash2 = async (req) => {
-  console.log('connect');
-
-  const file_owner = 'aljenfalkjwefnlakjwef';
-  // 현재 시크릿 키가 없어 사용 불가.
-  /* const im = await me(req);
-if (im.statusCode === 401 || im.statusCode === 500) {
-  return im;
-}
-const file_owner = im.userId;
-*/
-
-  try {
-    AWS.config.update(awsConfig);
-    const docClient = new AWS.DynamoDB.DocumentClient();
-
-    const parentOwnerParam = {
-      TableName: 'FileDirTable',
-      Key: {
-        file_owner: file_owner,
-        id: req.body.id,
-      },
-    };
-
-    const owner = (await docClient.get(parentOwnerParam).promise()).Item;
-
-    // 루트폴더의 주인인지 확인
-    if (file_owner !== owner.file_owner) {
-      return {
-        statusCode: 400,
-        success: false,
-        msg: '조회 권한이 없습니다.',
-      };
-    }
-
-    // 삭제된 폴더를 찾기위해 일단 user가 가지고있는 것들을 가져온다.
-    const rootValidParam = {
-      TableName: 'FileDirTable',
-      KeyConditionExpression: '#file_owner = :file_owner',
-      ExpressionAttributeNames: {
-        '#file_owner': 'file_owner',
-      },
-      ExpressionAttributeValues: {
-        ':file_owner': file_owner,
-      },
-    };
-
-    const folders = (await docClient.query(rootValidParam).promise()).Items;
-
-    let data = [];
-
-    for (let i = 0; i < folders.length; i++) {
-      if (folders[i].is_deleted === true) {
-        data.push(folders[i]);
-      }
-    }
-
-    console.log(JSON.stringify(data, null, 2));
-
-    const resMessage = {
-      statusCode: 200,
-      success: true,
-      msg: '휴지통 조회 완료',
-      data: data,
-    };
-    return resMessage;
-  } catch (err) {
-    console.log(err);
-    throw Error(err);
-  }
-};
